[Dap.Ipod] Cast SampleRate to ushort for older Mono

+++ src/Dap/Banshee.Dap.Ipod/Banshee.Dap.Ipod/IpodTrackInfo.cs
@@ -201,7 +201,7 @@ namespace Banshee.Dap.Ipod
 
             track.AlbumArtist = AlbumArtist;
             track.BitRate = BitRate;
-            track.SampleRate = SampleRate;
+            track.SampleRate = (ushort)SampleRate;
             track.BPM = (short)Bpm;
             track.Comment = Comment;
             track.Composer = Composer;
